# apps/api/openapi.yaml
openapi: 3.0.0
info:
  title: Formizee API
  description: The Forms Backend Platform API
  termsOfService: https://www.formizee.com/legal/terms-of-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    email: support@example.com
  version: 0.1.0
servers:
  - url: https://api.formizee.com/v1
    description: Production API
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    endpoint:
      type: object
      properties:
        id:
          type: string
          format: uuid

        name:
          type: string

        owner:
          type: string
          format: uuid

        isEnabled:
          type: boolean

        emailNotifications:
          type: boolean

        targetEmail:
          type: string
          format: email

        redirectUrl:
          type: string
          format: uri
    submission:
      type: object
      properties:
        id:
          type: integer

        endpoint:
          type: string
          format: uuid

        data:
          type: string
          format: binary

        files:
          type: array
          items:
            type: string
            format: uri

        isSpam:
          type: boolean

        createdAt:
          type: string
          format: date-time
    error:
      type: object
      properties:
        name:
          type: string

        message:
          type: string
          format: string
security:
  - session: []
...
paths:
...
  /endpoints/{id}:
    get:
      summary: Get Endpoint
      description: Returns the endpoint specified on the id.
      operationId: loadEndpoint
      tags: [Endpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Endpoint Loaded Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoint'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Endpoint Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      summary: Update Endpoint
      description: Updates the endpoint specified on the id.
      operationId: updateEndpoint
      tags: [Endpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string

                isEnabled:
                  type: boolean

                emailNotifications:
                  type: boolean

                targetEmail:
                  type: string
                  format: email

                redirectUrl:
                  type: string
                  format: uri

      responses:
        '200':
          description: Endpoint Updated Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/endpoint'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Endpoint Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      summary: Delete Endpoint
      description: Deletes the endpoint specified on the id.
      operationId: deleteEndpoint
      tags: [Endpoints]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Endpoint Deleted Successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Endpoint Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /endpoints:
    get:
      summary: Get All Endpoints
      description: Returns all user owned endpoints.
      operationId: loadAllEndpoints
      tags: [Endpoints]
      responses:
        '200':
          description: Endpoints Loaded Successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/endpoint'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: No Endpoints For The Current User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      summary: Create Endpoint
      description: Creates a new endpoint with the given name.
      operationId: saveEndpoint
      tags: [Endpoints]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '200':
          description: Endpoint Saved Successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/endpoint'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
  /submissions/{endpointId}/{id}:
    get:
      summary: Get Submission
      description: Returns the submission specified on the id.
      operationId: loadSubmission
      tags: [Submissions]
      parameters:
        - name: endpointId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Submission Loaded Successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Submission Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    patch:
      summary: Update Submission Data
      description: Updates the submission specified on the id.
      operationId: updateSubmission
      tags: [Submissions]
      parameters:
        - name: endpointId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isSpam:
                  type: boolean
      responses:
        '200':
          description: Submission Updated Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/submission'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Submission Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    delete:
      summary: Delete Submission
      description: Deletes the submission specified on the id.
      operationId: deleteSubmission
      tags: [Submissions]
      parameters:
        - name: endpointId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Submission Deleted Successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Submission Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /submissions/{endpointId}:
    get:
      summary: Get All Submissions
      description: Returns all endpoint owned submissions.
      operationId: loadAllSubmissions
      tags: [Submissions]
      parameters:
        - name: endpointId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Submissions Loaded Successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/submission'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Endpoint Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /submissions:
    post:
      summary: Create Submission
      description: Creates a new submission with the given data.
      operationId: saveSubmission
      tags: [Submissions]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                endpoint:
                  type: string
                  format: uuid
                data:
                  type: object
                  format: json
                files:
                  type: array
                  items:
                    type: string
                    format: uri
              required: [endpoint, data]
      responses:
        '200':
          description: Submission Saved Successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/submission'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '401':
          description: Unauthorized Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        '404':
          description: Endpoint Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...